[IMP] core: move win32 to monkeypatches
