+++ odoo/tools/__init__.py
@@ -8,7 +8,6 @@ from . import constants
 from . import pdf
 from . import pycompat
 from . import template_inheritance
-from . import win32
 from .parse_version import parse_version
 from .barcode import check_barcode_encoding
 from .cache import ormcache, ormcache_context
